+++ Network_Beacon_nrf54/src/network.h
@@ -13,6 +13,6 @@ void network_init(void);
 void network_apply_command(uint8_t parameter, uint16_t value);
 int network_params_load(void);
 int network_params_save(void);
-uint8_t network_read_contact(uint8_t *buffer, uint16_t buffer_len);
+uint16_t network_read_contact(uint8_t *buffer, uint16_t buffer_len);
 
 #endif /* NETWORK_H */
